# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PalPanel API
  version: 1.0.0
  description: Stable HTTP contract for the PalPanel 1.x management API.
servers:
  - url: /api
security:
  - sessionCookie: []
  - developmentKey: []
...
paths:
...
  /security/paldefender/starter-gift/grants/{id}/retry:
    parameters: [*idParameter]
    post: {operationId: retryStarterGift, tags: [security], x-palpanel-permission: 'security:write', responses: *writeResponses}
  /security/paldefender/starter-gift/grants/{id}:
    parameters: [*idParameter]
    delete: {operationId: forgetStarterGift, tags: [security], x-palpanel-permission: 'security:write', responses: *writeResponses}
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: palpanel_session
      description: HttpOnly SameSite=Lax session cookie created by registration or login.
    developmentKey:
      type: http
      scheme: bearer
      bearerFormat: ppk_...
      description: 'Revocable development key supplied as Authorization: Bearer ppk_...'
...